# front/src/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Le Renard Étudiant — API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/parcours/{slug}": {
      "get": {
        "tags": [
          "parcours"
        ],
        "summary": "Lire Parcours",
        "operationId": "lire_parcours_api_parcours__slug__get",
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Slug"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ParcoursDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "EtapeLecture": {
        "properties": {
          "slug": {
            "type": "string",
            "title": "Slug"
          },
          "titre": {
            "type": "string",
            "title": "Titre"
          },
          "duree_minutes": {
            "type": "integer",
            "title": "Duree Minutes"
          },
          "terminee": {
            "type": "boolean",
            "title": "Terminee"
          }
        },
        "type": "object",
        "required": [
          "slug",
          "titre",
          "duree_minutes",
          "terminee"
        ],
        "title": "EtapeLecture"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "Niveau": {
        "type": "string",
        "enum": [
          "debutant",
          "intermediaire"
        ],
        "title": "Niveau"
      },
...
      "ParcoursDetail": {
        "properties": {
          "slug": {
            "type": "string",
            "title": "Slug"
          },
          "titre": {
            "type": "string",
            "title": "Titre"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "niveau": {
            "$ref": "#/components/schemas/Niveau"
          },
          "theme": {
            "$ref": "#/components/schemas/ThemeLecture"
          },
          "nb_lecons": {
            "type": "integer",
            "title": "Nb Lecons"
          },
          "nb_terminees": {
            "type": "integer",
            "title": "Nb Terminees"
          },
          "duree_minutes": {
            "type": "integer",
            "title": "Duree Minutes"
          },
          "lecons": {
            "items": {
              "$ref": "#/components/schemas/EtapeLecture"
            },
            "type": "array",
            "title": "Lecons"
          },
          "prochaine_lecon": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Prochaine Lecon"
          }
        },
        "type": "object",
        "required": [
          "slug",
          "titre",
          "description",
          "niveau",
          "theme",
          "nb_lecons",
          "nb_terminees",
          "duree_minutes",
          "lecons",
          "prochaine_lecon"
        ],
        "title": "ParcoursDetail"
      },
...
      "ThemeLecture": {
        "properties": {
          "slug": {
            "type": "string",
            "title": "Slug"
          },
          "nom": {
            "type": "string",
            "title": "Nom"
          }
        },
        "type": "object",
        "required": [
          "slug",
          "nom"
        ],
        "title": "ThemeLecture"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}